+++ apps/web/client/src/app/project/[id]/_components/right-click-menu/index.tsx
@@ -187,7 +187,7 @@ export const RightClickMenu = observer(({ children }: RightClickMenuProps) => {
             console.error('No oid found');
             return;
         }
-        editorEngine.code.viewCodeBlock(oid);
+        editorEngine.ide.viewCodeBlock(oid);
     }
 
     return (
